+++ src/tests/buffer_ops_rect.rs
@@ -129,7 +129,7 @@ fn buffer_ops_rect() {
         //====================================================================
         // Reset vec:
         buf.cmd().read(&mut vec).rect([0, 0, 0], [0, 0, 0], dims, row_pitch, slc_pitch,
-            row_pitch, slc_pitch).queue(proque.queue()).block(true).enq().unwrap();
+            row_pitch, slc_pitch).queue(proque.queue()).enq().unwrap();
 
         // Run kernel:
         kernel_add.enq().expect("[FIXME]: HANDLE ME!");
@@ -139,7 +139,7 @@ fn buffer_ops_rect() {
 
         // Read from the random region into our vec:
         buf.cmd().read(&mut vec).rect(buf_origin, vec_origin, read_region.clone(), row_pitch,
-            slc_pitch, row_pitch, slc_pitch).queue(proque.queue()).block(true).enq().unwrap();
+            slc_pitch, row_pitch, slc_pitch).queue(proque.queue()).enq().unwrap();
 
         // Verify:
         tests::verify_vec_rect(vec_origin, read_region, cur_val, old_val,
@@ -208,11 +208,11 @@ fn buffer_ops_rect() {
         for ele in vec.iter_mut() { *ele = nxt_val }
 
         // Write to the random region:
-        buf.cmd().write(&mut vec).rect(buf_origin, vec_origin, read_region.clone(), row_pitch,
-            slc_pitch, row_pitch, slc_pitch).queue(proque.queue()).block(false).enq().unwrap();
+        buf.cmd().write(&vec).rect(buf_origin, vec_origin, read_region.clone(), row_pitch,
+            slc_pitch, row_pitch, slc_pitch).queue(proque.queue()).enq().unwrap();
         // Read the entire buffer back into the vector:
         buf.cmd().read(&mut vec).rect([0, 0, 0], [0, 0, 0], dims, row_pitch, slc_pitch,
-            row_pitch, slc_pitch).queue(proque.queue()).block(true).enq().unwrap();
+            row_pitch, slc_pitch).queue(proque.queue()).enq().unwrap();
         // Verify that our random region was in fact written correctly:
         tests::verify_vec_rect(buf_origin, read_region, nxt_val, cur_val,
             dims, 1, &vec, ttl_runs, true).unwrap();
@@ -306,14 +306,14 @@ fn buffer_ops_rect() {
 
         // Write the source vec to the source buf:
         buf_src.cmd().write(&vec_src).rect([0, 0, 0], [0, 0, 0], dims, row_pitch,
-            slc_pitch, row_pitch, slc_pitch).queue(proque.queue()).block(true).enq().unwrap();
+            slc_pitch, row_pitch, slc_pitch).queue(proque.queue()).enq().unwrap();
 
         // Copy from the source buffer to the random region on the destination buffer:
         buf_src.cmd().copy(&buf_dst, None, None).rect(src_origin, dst_origin, read_region.clone(), row_pitch,
             slc_pitch, row_pitch, slc_pitch).queue(proque.queue()).enq().unwrap();
         // Read the entire destination buffer into the destination vec:
         buf_dst.cmd().read(&mut vec_dst).rect([0, 0, 0], [0, 0, 0], dims, row_pitch, slc_pitch,
-            row_pitch, slc_pitch).queue(proque.queue()).block(true).enq().unwrap();
+            row_pitch, slc_pitch).queue(proque.queue()).enq().unwrap();
         // Verify that our random region was in fact written correctly:
         tests::verify_vec_rect(dst_origin, read_region, nxt_val, cur_val,
             dims, 1, &vec_dst, ttl_runs, true).unwrap();